BookStore/requirements.txt
streamlit
pandas
grpcio
grpcio-tools
